Run each vector within a test vector file within its own Activity
Previously, if a vector failed, we'd have no idea which vector it
was that failed without manually setting breakpoints.

This new structure categorizes each vector within an XCTContext,
named with the vector's description.

This means that when a vector fails, we can look at the Xcode output and
determine exactly which vector failed.